+++ sheets/snippets/tests/SpreadsheetBatchUpdateValuesTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright 2022 Google Inc.
+ * Copyright 2018 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
- require 'src/sheets_batch_get_values.php';
+ require 'src/SpreadsheetBatchUpdateValues.php';
 
- class SpreadSheetBatchGetValuesTest extends \PHPUnit\Framework\TestCase
+ class SpreadsheetBatchUpdateValuesTest extends \PHPUnit\Framework\TestCase
  {
-    public function testSpreadsheetBatchGetValues()
+    public function testSpreadSheetBatchUpdateValues()
     {   
-        $spreadSheet = batchGetValues('1sN_EOj0aYp5hn9DeqSY72G7sKaFRg82CsMGnK_Tooa8', 'Sheet1!A1:B2');
-        $this->assertNotNull($spreadSheet, 'ID not returned.');
+        $spreadSheet = batchUpdateValues('1sN_EOj0aYp5hn9DeqSY72G7sKaFRg82CsMGnK_Tooa8','Sheet1!A1:B2',"RAW", []);
+        $this->assertNotNull($spreadSheet, 'No result returned.');
     }
  }
